refactor: remove the s8() any-cast shim and fix what it was hiding
lib/supabase/sprint-8.ts cast the Supabase client to
SupabaseClient<any, "public", any> so queries against tables missing from
db/types.ts would compile. All 16 tables it covered now exist in the
generated types, so the shim is removed and its 66 call sites across 24
files go back to the typed client.

Removing it surfaced a real production bug. /api/cron/post-valuation-nurture
selects and updates nurture_day7_at / nurture_day30_at on
valuation_requests, but 0027 had guarded that ALTER on a table named
valuation_inquiries which has never existed, so the columns were never
created. The SELECT returned 42703, supabase-js reported data as null,
`data ?? []` collapsed to an empty loop, and the route returned
{ok: true, day7: 0, day30: 0} on every run while sending no email. The
any-cast is why the compiler could not see it. Migration 0058 adds the
columns where the code actually uses them, with partial indexes matching
the query shape; the route's docstring now records the history.

The remaining six errors were type-inference limits, not runtime faults,
and each was checked against PostgREST before being cast:

  - pgvector columns are emitted as `string`; PostgREST accepts a JSON
    number array and casts it (verified: upsert 201, match_properties RPC
    200 returning a row). Affects embeddings-backfill and concierge.
  - supabase-js cannot infer the `alias:fk_column(...)` embed form and
    widens the result to SelectQueryError. The queries are correct
    (verified 200 with data). Affects the two meilisearch routes,
    syndication load, and referrals.
  - integrations.setIntegrationStatus takes config as
    Record<string, unknown> against a Json column; narrowed at the write
    rather than pushing Json onto callers.

RawReferralWithReferee was ReferralRow & {referee}, but the query selects
only nine columns and toReferralEntry reads only those, so it is now a
Pick of exactly what is fetched.

Typecheck passes.

+++ app/(admin)/admin/properties/[id]/_actions.ts
@@ -11,7 +11,6 @@ import {
   type PropertyCompliance,
 } from "@/lib/schemas/property";
 import { propertyUrl } from "@/lib/queries/properties";
-import { s8 } from "@/lib/supabase/sprint-8";
 import {
   evaluatePublishability,
   type PublishabilityResult,
@@ -95,10 +94,7 @@ export async function updateProperty(
     .eq("id", id)
     .maybeSingle();
 
-  // s8() widens the .from() type so the Sprint 8 columns
-  // (bazar_verified, advisor_note, featured_on_homepage) typecheck
-  // until db/types.ts is regenerated against migrations 0015–0026.
-  const { data, error } = await s8(supabase)
+  const { data, error } = await supabase
     .from("properties")
     .update(updateData)
     .eq("id", id)
