+++ app/api/cron/post-valuation-nurture/route.ts
@@ -5,11 +5,12 @@
  * T-7 days (and hasn't received a day-7 nurture), email the
  * valuationNurtureDay7Template. Same for T-30 with day30 template.
  *
- * Idempotency: nurture_day7_at / nurture_day30_at fields on
- * valuation_requests (added in migration 0027). Until the migration
- * applies, the s8() escape-hatch lets the update typecheck and the
- * column-not-yet-present case fails gracefully (single-day re-sends
- * possible during the gap).
+ * Idempotency: nurture_day7_at / nurture_day30_at on valuation_requests,
+ * added in migration 0058. 0027 had intended to add them, but guarded the
+ * ALTER on a table named `valuation_inquiries` that has never existed in
+ * this schema — so the columns were never created, this route's SELECT
+ * returned 42703, `data ?? []` collapsed to an empty loop, and the cron
+ * reported {ok: true, day7: 0, day30: 0} while sending nothing.
  */
 
 import { NextResponse, type NextRequest } from "next/server";
@@ -21,7 +22,6 @@ import {
   valuationNurtureDay7Template,
   valuationNurtureDay30Template,
 } from "@/lib/email-templates";
-import { s8 } from "@/lib/supabase/sprint-8";
 import type { Database } from "@/db/types";
 
 function adminClient() {
@@ -67,7 +67,7 @@ export async function GET(req: NextRequest) {
     // Day 7
     {
       const w = dayWindow(7);
-      const { data } = await s8(supabase)
+      const { data } = await supabase
         .from("valuation_requests")
         .select(
           "id, owner_name, owner_email, estimate_mid_aed, sent_at, nurture_day7_at",
@@ -90,7 +90,7 @@ export async function GET(req: NextRequest) {
         });
         if (sent.status === "ok") {
           day7 += 1;
-          await s8(supabase)
+          await supabase
             .from("valuation_requests")
             .update({ nurture_day7_at: new Date().toISOString() })
             .eq("id", r.id);
@@ -101,7 +101,7 @@ export async function GET(req: NextRequest) {
     // Day 30
     {
       const w = dayWindow(30);
-      const { data } = await s8(supabase)
+      const { data } = await supabase
         .from("valuation_requests")
         .select("id, owner_name, owner_email, sent_at, nurture_day30_at")
         .gte("sent_at", w.start)
@@ -121,7 +121,7 @@ export async function GET(req: NextRequest) {
         });
         if (sent.status === "ok") {
           day30 += 1;
-          await s8(supabase)
+          await supabase
             .from("valuation_requests")
             .update({ nurture_day30_at: new Date().toISOString() })
             .eq("id", r.id);
